Miscellaneous changes

+++ src/main/java/bookstore/inventory/ShoppingCart.java
@@ -23,7 +23,7 @@ public class ShoppingCart {
     private List<InventoryItem> booksInCart = new ArrayList<>();
     @ManyToOne (fetch = FetchType.EAGER)
     private Inventory inventory;
-    private float totalPrice;
+    private Double totalPrice;
 
     /**
      * Constructor for ShoppingCart
@@ -31,7 +31,7 @@ public class ShoppingCart {
      */
     public ShoppingCart(Inventory inventory){
         this.inventory = inventory;
-        this.totalPrice = 0;
+        this.totalPrice = 0.0;
     }
 
     /**
@@ -39,7 +39,7 @@ public class ShoppingCart {
      */
     public ShoppingCart() {
         this.inventory = new Inventory();
-        this.totalPrice = 0;
+        this.totalPrice = 0.0;
     }
 
     /**
@@ -133,7 +133,7 @@ public class ShoppingCart {
      * Method to update total price
      */
     private void updateTotalPrice(){
-        totalPrice = 0;
+        totalPrice = 0.0;
         //update price
         for (InventoryItem item : booksInCart){
             totalPrice += (float) (item.getBook().getPrice() * item.getQuantity());
@@ -144,7 +144,7 @@ public class ShoppingCart {
      * Method to get the total price of cart
      * @return  the total price
      */
-    public float getTotalPrice() {
+    public Double getTotalPrice() {
         return totalPrice;
     }
